fix: deep security + bugs + maintainability audit across all layers
Security: JwtFilter path traversal fix, input validation on all API
controllers, SQL injection prevention in BFF routes, onboarding
brute-force protection (8-hex codes), foto auth via header not URL.
Bugs: DAO tenant leaks, model null safety, SyncClient retry hardening.
Maintainability: GUI controllers refactored, model consistency.
OCR: fetchFoto via Authorization header, camera error handling,
offline queue race condition fix, Gemini response parsing.

+++ src/model/EncomendaFinanceiro.java
@@ -30,6 +30,6 @@ public class EncomendaFinanceiro {
     public String getRestanteFormatado() { return String.format("R$ %,.2f", getRestante()); }
 
     public String getStatus() {
-        return StatusPagamento.calcularPorSaldo(getRestante().doubleValue(), getPago().doubleValue()).name();
+        return StatusPagamento.calcularPorSaldo(getRestante(), getPago()).name();
     }
 }
